JobGad/Backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
python-multipart>=0.0.9

# Database & ORM
SQLAlchemy>=2.0.30
alembic>=1.13.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.3

# AI & Agent Framework
crewai>=0.35.0
langchain>=0.2.6
langchain-google-genai>=1.0.6
google-generativeai>=0.5.4

# Embeddings & Vector Database
pinecone-client>=4.1.1
sentence-transformers>=3.0.1

# Document Processing
PyMuPDF>=1.24.5        # Good for PDF parsing (alternative to PyPDF2)
pdfplumber>=0.11.1
python-docx>=1.1.2

# Real-time WebSockets
python-socketio>=5.11.3
websockets>=12.0

# Async Tasks
celery>=5.4.0
redis>=5.0.7

# Integrations & Utilities
PyGithub>=2.3.0
beautifulsoup4>=4.12.3
requests>=2.32.3
httpx>=0.27.0
python-magic>=0.4.27; sys_platform != "win32"
python-magic-bin>=0.4.14; sys_platform == "win32"
python-multipart>=0.0.9
slowapi>=0.1.9
supabase>=2.4.0

# AI-powered Search
tavily-python>=0.3.3